add abstract link parser
